+++ libsolidity/formal/BMC.h
@@ -62,7 +62,7 @@ public:
 		langutil::UniqueErrorReporter& _errorReporter,
 		std::map<h256, std::string> const& _smtlib2Responses,
 		ReadCallback::Callback const& _smtCallback,
-		ModelCheckerSettings const& _settings,
+		ModelCheckerSettings _settings,
 		langutil::CharStreamProvider const& _charStreamProvider
 	);