+++ libsolidity/formal/ModelChecker.cpp
@@ -20,6 +20,13 @@
 #ifdef HAVE_Z3
 #include <libsmtutil/Z3Interface.h>
 #endif
+#ifdef HAVE_Z3_DLOPEN
+#include <z3_version.h>
+#endif
+
+#if defined(__linux) || defined(__APPLE__)
+#include <boost/process.hpp>
+#endif
 
 #include <range/v3/algorithm/any_of.hpp>
 #include <range/v3/view.hpp>
@@ -29,6 +36,7 @@ using namespace solidity;
 using namespace solidity::util;
 using namespace solidity::langutil;
 using namespace solidity::frontend;
+using namespace solidity::smtutil;
 
 ModelChecker::ModelChecker(
 	ErrorReporter& _errorReporter,
@@ -46,13 +54,11 @@ ModelChecker::ModelChecker(
 }
 
 // TODO This should be removed for 0.9.0.
-void ModelChecker::enableAllEnginesIfPragmaPresent(vector<shared_ptr<SourceUnit>> const& _sources)
+bool ModelChecker::isPragmaPresent(vector<shared_ptr<SourceUnit>> const& _sources)
 {
-	bool hasPragma = ranges::any_of(_sources, [](auto _source) {
+	return ranges::any_of(_sources, [](auto _source) {
 		return _source && _source->annotation().experimentalFeatures.count(ExperimentalFeature::SMTChecker);
 	});
-	if (hasPragma)
-		m_settings.engine = ModelCheckerEngine::All();
 }
 
 void ModelChecker::checkRequestedSourcesAndContracts(vector<shared_ptr<SourceUnit>> const& _sources)
@@ -135,9 +141,12 @@ vector<string> ModelChecker::unhandledQueries()
 	return m_bmc.unhandledQueries() + m_chc.unhandledQueries();
 }
 
-solidity::smtutil::SMTSolverChoice ModelChecker::availableSolvers()
+SMTSolverChoice ModelChecker::availableSolvers()
 {
 	smtutil::SMTSolverChoice available = smtutil::SMTSolverChoice::SMTLIB2();
+#if defined(__linux) || defined(__APPLE__)
+	available.eld = !boost::process::search_path("eld").empty();
+#endif
 #ifdef HAVE_Z3
 	available.z3 = solidity::smtutil::Z3Interface::available();
 #endif
@@ -145,4 +154,48 @@ solidity::smtutil::SMTSolverChoice ModelChecker::availableSolvers()
 	available.cvc4 = true;
 #endif
 	return available;
+}
+
+SMTSolverChoice ModelChecker::checkRequestedSolvers(SMTSolverChoice _enabled, ErrorReporter& _errorReporter)
+{
+	SMTSolverChoice availableSolvers{ModelChecker::availableSolvers()};
+
+	if (_enabled.cvc4 && !availableSolvers.cvc4)
+	{
+		_enabled.cvc4 = false;
+		_errorReporter.warning(
+			4902_error,
+			SourceLocation(),
+			"Solver CVC4 was selected for SMTChecker but it is not available."
+		);
+	}
+
+	if (_enabled.eld && !availableSolvers.eld)
+	{
+		_enabled.eld = false;
+		_errorReporter.warning(
+			4458_error,
+			SourceLocation(),
+#if defined(__linux) || defined(__APPLE__)
+			"Solver Eldarica was selected for SMTChecker but it was not found in the system."
+#else
+			"Solver Eldarica was selected for SMTChecker but it is only supported on Linux and MacOS."
+#endif
+		);
+	}
+
+	if (_enabled.z3 && !availableSolvers.z3)
+	{
+		_enabled.z3 = false;
+		_errorReporter.warning(
+			8158_error,
+			SourceLocation(),
+			"Solver z3 was selected for SMTChecker but it is not available."
+#ifdef HAVE_Z3_DLOPEN
+			" libz3.so." + to_string(Z3_MAJOR_VERSION) + "." + to_string(Z3_MINOR_VERSION) + " was not found."
+#endif
+		);
+	}
+
+	return _enabled;
 }