+++ libsolidity/formal/ModelChecker.h
@@ -58,7 +58,7 @@ public:
 	);
 
 	// TODO This should be removed for 0.9.0.
-	void enableAllEnginesIfPragmaPresent(std::vector<std::shared_ptr<SourceUnit>> const& _sources);
+	static bool isPragmaPresent(std::vector<std::shared_ptr<SourceUnit>> const& _sources);
 
 	/// Generates error messages if the requested sources and contracts
 	/// do not exist.
@@ -73,6 +73,10 @@ public:
 
 	/// @returns SMT solvers that are available via the C++ API.
 	static smtutil::SMTSolverChoice availableSolvers();
+
+	/// @returns the intersection of the enabled and available solvers,
+	/// reporting warnings when a solver is enabled but not available.
+	static smtutil::SMTSolverChoice checkRequestedSolvers(smtutil::SMTSolverChoice _enabled, langutil::ErrorReporter& _errorReporter);
 
 private:
 	/// Error reporter from CompilerStack.