+++ libsolidity/formal/SMTEncoder.cpp
@@ -47,13 +47,13 @@ using namespace solidity::frontend;
 
 SMTEncoder::SMTEncoder(
 	smt::EncodingContext& _context,
-	ModelCheckerSettings const& _settings,
+	ModelCheckerSettings _settings,
 	UniqueErrorReporter& _errorReporter,
 	langutil::CharStreamProvider const& _charStreamProvider
 ):
 	m_errorReporter(_errorReporter),
 	m_context(_context),
-	m_settings(_settings),
+	m_settings(std::move(_settings)),
 	m_charStreamProvider(_charStreamProvider)
 {
 }