+++ libsolidity/formal/SMTEncoder.h
@@ -54,7 +54,7 @@ class SMTEncoder: public ASTConstVisitor
 public:
 	SMTEncoder(
 		smt::EncodingContext& _context,
-		ModelCheckerSettings const& _settings,
+		ModelCheckerSettings _settings,
 		langutil::UniqueErrorReporter& _errorReporter,
 		langutil::CharStreamProvider const& _charStreamProvider
 	);
@@ -483,7 +483,7 @@ protected:
 	/// Stores the context of the encoding.
 	smt::EncodingContext& m_context;
 
-	ModelCheckerSettings const& m_settings;
+	ModelCheckerSettings m_settings;
 
 	/// Character stream for each source,
 	/// used for retrieving source text of expressions for e.g. counter-examples.