+++ libsolutil/TemporaryDirectory.cpp
@@ -16,9 +16,9 @@
 */
 // SPDX-License-Identifier: GPL-3.0
 
-#include <test/TemporaryDirectory.h>
+#include <libsolutil/TemporaryDirectory.h>
 
-#include <test/libsolidity/util/SoltestErrors.h>
+#include <liblangutil/Exceptions.h>
 
 #include <boost/algorithm/string/predicate.hpp>
 #include <boost/filesystem.hpp>
@@ -28,7 +28,7 @@
 
 using namespace std;
 using namespace solidity;
-using namespace solidity::test;
+using namespace solidity::util;
 
 namespace fs = boost::filesystem;
 
@@ -36,7 +36,7 @@ TemporaryDirectory::TemporaryDirectory(std::string const& _prefix):
 	m_path(fs::temp_directory_path() / fs::unique_path(_prefix + "-%%%%-%%%%-%%%%-%%%%"))
 {
 	// Prefix should just be a file name and not contain anything that would make us step out of /tmp.
-	soltestAssert(fs::path(_prefix) == fs::path(_prefix).stem(), "");
+	solAssert(fs::path(_prefix) == fs::path(_prefix).stem(), "");
 
 	fs::create_directory(m_path);
 }
@@ -49,8 +49,8 @@ TemporaryDirectory::TemporaryDirectory(
 {
 	for (boost::filesystem::path const& subdirectory: _subdirectories)
 	{
-		soltestAssert(!subdirectory.is_absolute() && subdirectory.root_path() != "/", "");
-		soltestAssert(
+		solAssert(!subdirectory.is_absolute() && subdirectory.root_path() != "/", "");
+		solAssert(
 			m_path.lexically_relative(subdirectory).empty() ||
 			*m_path.lexically_relative(subdirectory).begin() != "..",
 			""
@@ -62,10 +62,10 @@ TemporaryDirectory::TemporaryDirectory(
 TemporaryDirectory::~TemporaryDirectory()
 {
 	// A few paranoid sanity checks just to be extra sure we're not deleting someone's homework.
-	soltestAssert(m_path.string().find(fs::temp_directory_path().string()) == 0, "");
-	soltestAssert(!fs::equivalent(m_path, fs::temp_directory_path()), "");
-	soltestAssert(!fs::equivalent(m_path, m_path.root_path()), "");
-	soltestAssert(!m_path.empty(), "");
+	solAssert(m_path.string().find(fs::temp_directory_path().string()) == 0, "");
+	solAssert(!fs::equivalent(m_path, fs::temp_directory_path()), "");
+	solAssert(!fs::equivalent(m_path, m_path.root_path()), "");
+	solAssert(!m_path.empty(), "");
 
 	boost::system::error_code errorCode;
 	uintmax_t numRemoved = fs::remove_all(m_path, errorCode);